Changed totalTicks to be of type Long
Also corrected error in totalTicks in BreakoutGameState (was using CaveSwing total by mistake)

+++ src/games/caveswing/CaveSwing.kt
@@ -91,10 +91,10 @@ class CaveGameState : ExtendedAbstractGameState, Serializable {
     }
 
     companion object {
-        var totalTicks: Int = 0
+        var totalTicks: Long = 0
     }
 
-    override fun totalTicks(): Int {
+    override fun totalTicks(): Long {
         return totalTicks
     }
